requirements.txt
Django>=5.2.5,<6.0
djangorestframework>=3.15.0,<4.0
django-filter>=24.3,<25.0
django-cors-headers>=4.6.0,<5.0
djangorestframework-simplejwt>=5.3.1,<6.0
python-decouple>=3.8,<4.0
celery>=5.4.0,<6.0
kombu>=5.4.0,<6.0
billiard>=4.2.0,<5.0

